+++ src/Stack/Options/BuildMonoidParser.hs
@@ -74,9 +74,10 @@ buildOptsMonoidParser hide0 =
             hide)
     options =
         BuildOptsMonoid <$> libProfiling <*> exeProfiling <*> haddock <*>
-        haddockOptsParser hide0 <*> openHaddocks <*>
-        haddockDeps <*> copyBins <*> preFetch <*> keepGoing <*> forceDirty <*>
-        tests <*> testOptsParser hide0 <*> benches <*> benchOptsParser hide0 <*> reconfigure <*>
+        haddockOptsParser hide0 <*> openHaddocks <*> haddockDeps <*>
+        haddockInternal <*> copyBins <*> preFetch <*> keepGoing <*>
+        forceDirty <*> tests <*> testOptsParser hide0 <*> benches <*>
+        benchOptsParser hide0 <*> reconfigure <*>
         cabalVerbose <*> splitObjs
     libProfiling =
         firstBoolFlags
@@ -100,6 +101,11 @@ buildOptsMonoidParser hide0 =
             hide
     haddockDeps =
         firstBoolFlags "haddock-deps" "building Haddocks for dependencies" hide
+    haddockInternal =
+        firstBoolFlags
+            "haddock-internal"
+            "building Haddocks for internal modules (like cabal haddock --internal)"
+            hide
     copyBins =
         firstBoolFlags
             "copy-bins"